[time]change struct timeval type for Y2038 prebrom

+++ src/lib/libc/time/__secs_to_tm.c
@@ -11,9 +11,9 @@
 #define DAYS_PER_100Y	(365 * 100 + 24)
 #define DAYS_PER_4Y		(365 * 4 + 1)
 
-int __secs_to_tm(long long t, struct tm * tm)
+int __secs_to_tm(int64_t t, struct tm * tm)
 {
-	long long days, secs, years;
+	int64_t days, secs, years;
 	int remdays, remsecs, remyears;
 	int qc_cycles, c_cycles, q_cycles;
 	int months;
